Cria a classe RedisQueue e a interface Queue, refatorando a lógica de gerenciamento de filas
